Fixed endian conversion code to be more portable / efficient / non-UB

+++ ReaK/core/serialization/bin_archiver.h
@@ -70,14 +70,14 @@ class bin_iarchive : public iarchive {
   iarchive& load_unsigned_char(
       const std::pair<std::string, unsigned char&>& u) override;
 
-  iarchive& load_int(std::ptrdiff_t& i) override;
+  iarchive& load_int(std::int64_t& i) override;
 
-  iarchive& load_int(const std::pair<std::string, std::ptrdiff_t&>& i) override;
+  iarchive& load_int(const std::pair<std::string, std::int64_t&>& i) override;
 
-  iarchive& load_unsigned_int(std::size_t& u) override;
+  iarchive& load_unsigned_int(std::uint64_t& u) override;
 
   iarchive& load_unsigned_int(
-      const std::pair<std::string, std::size_t&>& u) override;
+      const std::pair<std::string, std::uint64_t&>& u) override;
 
   iarchive& load_float(float& f) override;
 
@@ -137,14 +137,14 @@ class bin_oarchive : public oarchive {
   oarchive& save_unsigned_char(
       const std::pair<std::string, unsigned char>& u) override;
 
-  oarchive& save_int(std::ptrdiff_t i) override;
+  oarchive& save_int(std::int64_t i) override;
 
-  oarchive& save_int(const std::pair<std::string, std::ptrdiff_t>& i) override;
+  oarchive& save_int(const std::pair<std::string, std::int64_t>& i) override;
 
-  oarchive& save_unsigned_int(std::size_t u) override;
+  oarchive& save_unsigned_int(std::uint64_t u) override;
 
   oarchive& save_unsigned_int(
-      const std::pair<std::string, std::size_t>& u) override;
+      const std::pair<std::string, std::uint64_t>& u) override;
 
   oarchive& save_float(float f) override;
